Move `printDecorators` into `class.js`
